requirements.txt
discord.py
asyncpg
tqdm
langchain
langchain_community
sentence-transformers
requests